Backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database & ORM
supabase==2.3.4
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.9.2
pydantic-settings==2.5.0
email-validator==2.1.0

# Communication Integrations
resend==0.7.0
twilio==8.11.1
sendgrid==6.11.0

# Task Queue & Background Jobs
celery==5.3.6
redis==5.0.1

# Utilities
httpx==0.26.0
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3

# Monitoring & Logging
structlog==24.1.0
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
